refactor(spotify): refactor spotify command and related services
This commit refactors the spotify command and related services to improve code organization and readability. The code has been moved from the spotify.js file to separate service files, such as spotify.js, message.js, and audio.js. The spotify command now uses the spotify-web-api-node library to search for songs and download audio files. The refreshAccessToken function has been added to handle refreshing the access token for the Spotify API. The searchAndDownloadSong function has been implemented to search for songs on Spotify and download the audio file. The sendAudio function has been added to send the downloaded audio file to the user. The sendMessage function has been added to send messages to the user. Overall, these changes enhance the functionality and maintainability of the spotify command.